Fix jasmine shim again.
This was introduced in 6fefb.
